Fix numpy float issues

+++ insightface_func/utils/face_align_ffhqandnewarc.py
@@ -13,26 +13,26 @@ from skimage import transform as trans
 
 src1 = np.array([[51.642, 50.115], [57.617, 49.990], [35.740, 69.007],
                  [51.157, 89.050], [57.025, 89.702]],
-                dtype=np.float32)
+                dtype=float)
 #<--left
 src2 = np.array([[45.031, 50.118], [65.568, 50.872], [39.677, 68.111],
                  [45.177, 86.190], [64.246, 86.758]],
-                dtype=np.float32)
+                dtype=float)
 
 #---frontal
 src3 = np.array([[39.730, 51.138], [72.270, 51.138], [56.000, 68.493],
                  [42.463, 87.010], [69.537, 87.010]],
-                dtype=np.float32)
+                dtype=float)
 
 #-->right
 src4 = np.array([[46.845, 50.872], [67.382, 50.118], [72.737, 68.111],
                  [48.167, 86.758], [67.236, 86.190]],
-                dtype=np.float32)
+                dtype=float)
 
 #-->right profile
 src5 = np.array([[54.796, 49.990], [60.771, 50.115], [76.673, 69.007],
                  [55.388, 89.702], [61.257, 89.050]],
-                dtype=np.float32)
+                dtype=float)
 
 src = np.array([src1, src2, src3, src4, src5])
 src_map = src
@@ -44,7 +44,7 @@ ffhq_src = np.expand_dims(ffhq_src, axis=0)
 # arcface_src = np.array(
 #     [[38.2946, 51.6963], [73.5318, 51.5014], [56.0252, 71.7366],
 #      [41.5493, 92.3655], [70.7299, 92.2041]],
-#     dtype=np.float32)
+#     dtype=float)
 
 # arcface_src = np.expand_dims(arcface_src, axis=0)
 
@@ -125,10 +125,10 @@ def transform(data, center, output_size, scale, rotation):
 
 
 def trans_points2d(pts, M):
-    new_pts = np.zeros(shape=pts.shape, dtype=np.float32)
+    new_pts = np.zeros(shape=pts.shape, dtype=float)
     for i in range(pts.shape[0]):
         pt = pts[i]
-        new_pt = np.array([pt[0], pt[1], 1.], dtype=np.float32)
+        new_pt = np.array([pt[0], pt[1], 1.], dtype=float)
         new_pt = np.dot(M, new_pt)
         #print('new_pt', new_pt.shape, new_pt)
         new_pts[i] = new_pt[0:2]
@@ -139,10 +139,10 @@ def trans_points2d(pts, M):
 def trans_points3d(pts, M):
     scale = np.sqrt(M[0][0] * M[0][0] + M[0][1] * M[0][1])
     #print(scale)
-    new_pts = np.zeros(shape=pts.shape, dtype=np.float32)
+    new_pts = np.zeros(shape=pts.shape, dtype=float)
     for i in range(pts.shape[0]):
         pt = pts[i]
-        new_pt = np.array([pt[0], pt[1], 1.], dtype=np.float32)
+        new_pt = np.array([pt[0], pt[1], 1.], dtype=float)
         new_pt = np.dot(M, new_pt)
         #print('new_pt', new_pt.shape, new_pt)
         new_pts[i][0:2] = new_pt[0:2]
